Проведен рефакторинг контроллеров. Для контроллеров уменьшена связанность Dependency Inversion, также логика игры отделена от UI.
