+++ DropHub/Controllers/Base.lproj/LaunchScreen.storyboard
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="24128" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" launchScreen="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="01J-lp-oVM">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="24063"/>
+        <capability name="Named colors" minToolsVersion="9.0"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--View Controller-->
+        <scene sceneID="EHf-IW-A2E">
+            <objects>
+                <viewController id="01J-lp-oVM" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="Ze5-6b-2t3">
+                        <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="USP_7" translatesAutoresizingMaskIntoConstraints="NO" id="9bl-q0-RCi">
+                                <rect key="frame" x="31.666666666666657" y="339" width="330" height="224"/>
+                                <constraints>
+                                    <constraint firstAttribute="width" constant="330" id="ukf-NU-rO9"/>
+                                </constraints>
+                            </imageView>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Drop" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="D0P-zT-dCU">
+                                <rect key="frame" x="47" y="418.66666666666669" width="175" height="94.333333333333314"/>
+                                <fontDescription key="fontDescription" name="Thonburi-Bold" family="Thonburi" pointSize="72"/>
+                                <color key="textColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                <color key="highlightedColor" red="0.470796459" green="0.86078912019999998" blue="0.52561910000000001" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
+                            </label>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" misplaced="YES" text="HUB" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="BTY-HV-BMV">
+                                <rect key="frame" x="230" y="473" width="63" height="40"/>
+                                <constraints>
+                                    <constraint firstAttribute="height" constant="40" id="yQ5-Qi-5Nq"/>
+                                </constraints>
+                                <fontDescription key="fontDescription" name="Thonburi-Bold" family="Thonburi" pointSize="27"/>
+                                <color key="textColor" name="imageBackColor"/>
+                                <color key="highlightedColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                            </label>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
+                        <color key="backgroundColor" name="firstGreen"/>
+                        <constraints>
+                            <constraint firstItem="9bl-q0-RCi" firstAttribute="centerY" secondItem="6Tk-OE-BBY" secondAttribute="centerY" id="8wJ-7m-O6n"/>
+                            <constraint firstItem="D0P-zT-dCU" firstAttribute="bottom" secondItem="9bl-q0-RCi" secondAttribute="bottom" constant="-50" id="JT9-JW-I4m"/>
+                            <constraint firstItem="BTY-HV-BMV" firstAttribute="top" secondItem="9bl-q0-RCi" secondAttribute="top" constant="134" id="Ov0-ry-hxK"/>
+                            <constraint firstItem="D0P-zT-dCU" firstAttribute="centerX" secondItem="9bl-q0-RCi" secondAttribute="centerX" constant="-62.166666666666657" id="QDG-nr-XQq"/>
+                            <constraint firstItem="BTY-HV-BMV" firstAttribute="leading" secondItem="D0P-zT-dCU" secondAttribute="trailing" constant="8" id="ReC-0a-kHz"/>
+                            <constraint firstItem="D0P-zT-dCU" firstAttribute="leading" secondItem="9bl-q0-RCi" secondAttribute="leading" constant="15.333333333333343" id="Rn7-xH-XaO"/>
+                            <constraint firstItem="9bl-q0-RCi" firstAttribute="height" secondItem="Ze5-6b-2t3" secondAttribute="height" multiplier="0.262911" id="bYs-7K-4gR"/>
+                            <constraint firstItem="9bl-q0-RCi" firstAttribute="centerX" secondItem="6Tk-OE-BBY" secondAttribute="centerX" id="mrc-be-84X"/>
+                        </constraints>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="iYj-Kq-Ea1" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="52.671755725190835" y="374.64788732394368"/>
+        </scene>
+    </scenes>
+    <resources>
+        <image name="USP_7" width="1858" height="798"/>
+        <namedColor name="firstGreen">
+            <color red="0.20000000000000001" green="0.41960784313725491" blue="0.34509803921568627" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
+        </namedColor>
+        <namedColor name="imageBackColor">
+            <color red="0.0078431372549019607" green="0.12156862745098039" blue="0.090196078431372548" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
+        </namedColor>
+    </resources>
+</document>
