+++ DropHub/Controllers/Protocols/FullImageControllerProtocol.swift
@@ -1,0 +1,7 @@
+//  FullImageControllerProtocol.swift
+
+import UIKit
+
+protocol FullImageControllerProtocol: AnyObject {
+    func setImage(_ image: UIImage)
+}
