+++ DropHub/Controllers/Protocols/GameDropAlgorithmProtocol.swift
@@ -1,0 +1,5 @@
+//  GameDropAlgorithmProtocol.swift
+
+protocol GameDropAlgorithmProtocol: AnyObject {
+    func startGame()
+}
