+++ DropHub/Controllers/Protocols/GameLayoutProtocol.swift
@@ -1,0 +1,6 @@
+//  GameLayoutProtocol.swift
+
+protocol GameLayoutProtocol: AnyObject {
+    func changeMainImageTopConstraint()
+    func changeMainImageLeftAndRightConstraint()
+}
