+++ DropHub/Controllers/Protocols/GameViewControllerPresentationProtocol.swift
@@ -1,0 +1,6 @@
+//  GameViewControllerPresentationProtocol.swift
+
+protocol GameViewControllerPresentationProtocol: AnyObject {
+    func showAlert()
+    func presentTextLabel()
+}
